torch_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
anyio=3.5.0=py310h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
atk-1.0=2.36.0=ha1a6a79_0
attrs=22.1.0=py310h06a4308_0
autopep8=1.6.0=pyhd3eb1b0_1
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310h06a4308_0
blas=2.106=openblas
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
cairo=1.16.0=hb05425b_3
certifi=2022.9.24=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.0.5=py310hdb19cb5_0
cryptography=38.0.1=py310h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
findspark=2.0.1=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.1=hef1e5e3_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fribidi=1.0.10=h7b6447c_0
gdk-pixbuf=2.42.8=h433bba3_1
giflib=5.2.1=h7b6447c_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gobject-introspection=1.72.0=py310hbb6d50b_1
graphite2=1.3.14=h295c915_1
graphviz=2.50.0=h3cd0ef9_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gtk2=2.24.33=h73c1081_2
gts=0.7.6=hb67d8dd_3
h5py=3.7.0=py310he06866b_0
harfbuzz=4.3.0=hd55b92a_0
hdf5=1.10.6=h3ffc7dd_1
icu=58.2=he6710b0_3
idna=3.4=py310h06a4308_0
intel-openmp=2022.1.0=h9e868ea_3769
ipykernel=6.15.2=py310h06a4308_0
ipython=8.7.0=py310h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py310h06a4308_1
jinja2=3.1.2=py310h06a4308_0
jpeg=9e=h7f8727e_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.16.0=py310h06a4308_0
jupyter=1.0.0=py310h06a4308_8
jupyter-resource-usage=0.6.4=pyhd8ed1ab_0
jupyter_client=7.4.7=py310h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_contrib_core=0.4.0=pyhd8ed1ab_0
jupyter_contrib_nbextensions=0.5.1=pyhd8ed1ab_2
jupyter_core=4.11.2=py310h06a4308_0
jupyter_highlight_selected_word=0.2.0=py310hff52083_1005
jupyter_latex_envs=1.4.6=pyhd8ed1ab_1002
jupyter_nbextensions_configurator=0.4.1=pyhd8ed1ab_2
jupyter_server=1.18.1=py310h06a4308_0
jupyterlab=3.5.0=py310h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.16.3=py310h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py310h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=6_openblas
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcblas=3.9.0=6_openblas
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=12.2.0=h65d4601_19
libgd=2.3.3=h695aa2c_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=6_openblas
liblapacke=3.9.0=6_openblas
libllvm10=10.0.1=hbcb73fb_5
libllvm14=14.0.6=h9e868ea_1
libopenblas=0.3.12=pthreads_h4812303_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
librsvg=2.54.4=h19fe530_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=12.2.0=h46fd767_19
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_2
libtool=2.4.6=h6a678d5_1009
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
llvm=14.0.6=h5c3875c_1
llvm-openmp=14.0.6=h9e868ea_0
lxml=4.9.1=py310h1edc446_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=py310h7f8727e_0
matplotlib-base=3.6.2=py310h945d387_0
matplotlib-inline=0.1.6=py310h06a4308_0
mistune=0.8.4=py310h7f8727e_1000
mkl=2022.1.0=hc2b9512_224
munkres=1.1.4=py_0
nb_conda_kernels=2.3.1=py310h06a4308_0
nbclassic=0.4.8=py310h06a4308_0
nbclient=0.5.13=py310h06a4308_0
nbconvert=6.5.4=py310h06a4308_0
nbformat=5.7.0=py310h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nomkl=3.0=0
notebook=6.5.2=py310h06a4308_0
notebook-shim=0.2.2=py310h06a4308_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.3=py310hf05e7a9_101
numpy=1.23.5=py310h53a5b5f_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py310h6a678d5_0
pandocfilters=1.5.0=pyhd3eb1b0_0
pango=1.50.7=h05da053_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hace64e9_1
pip=22.3.1=py310h06a4308_0
pixman=0.40.0=h7f8727e_1
ply=3.11=py310h06a4308_0
portalocker=2.3.0=py310h06a4308_0
prometheus_client=0.14.1=py310h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py4j=0.10.9.7=pypi_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pyrsistent=0.18.0=py310h7f8727e_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.8=h7a1cb2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py310h06a4308_0
python-graphviz=0.20.1=py310h06a4308_0
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py310h06a4308_0
pyyaml=6.0=py310h5eee18b_1
pyzmq=23.2.0=py310h6a678d5_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtconsole=5.3.2=py310h06a4308_0
qtpy=2.2.0=py310h06a4308_0
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=py310h06a4308_0
seaborn=0.12.0=py310h06a4308_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.5.0=py310h06a4308_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.5.1=h9b69904_2
sniffio=1.2.0=py310h06a4308_1
soupsieve=2.3.2.post1=py310h06a4308_0
sqlite=3.40.0=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py310h06a4308_0
tinycss2=1.2.1=py310h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py310h06a4308_0
torchaudio=0.12.1=py310_cu113
torchdata=0.4.1=pyh8b8bddf_0
torchmetrics=0.11.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.13.1=cpu_py310h164cc8f_0
torchviz=0.0.2=pypi_0
tornado=6.2=py310h5eee18b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.4.0=py310h06a4308_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=py310h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310h06a4308_1
websocket-client=0.58.0=py310h06a4308_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py310h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
